% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attackerkb-package.R
\docType{package}
\name{attackerkb}
\alias{attackerkb}
\alias{attackerkb-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/attackerkb}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/attackerkb/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
