% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{attackerkb_api_key}
\alias{attackerkb_api_key}
\title{Get or set ATTACKERKB_API_KEY value}
\usage{
attackerkb_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new AttackerKB key for the current environment?}
}
\value{
atomic character vector containing the AttackerKB API key
}
\description{
The API wrapper functions in this package all rely on a AttackerKB API
key residing in the environment variable \code{ATTACKERKB_API_KEY}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
\details{
You can obtain an AttackerKB API key by going visiting
\strong{AttackerKB Profile Page > Settings > API Key.}
}
