% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{kb_assessments}
\alias{kb_assessments}
\alias{kb_assessment}
\title{Helpers to query AttackerKB assessments}
\usage{
kb_assessments(
  assessment_id = NULL,
  editor_id = NULL,
  topid_id = NULL,
  created = NULL,
  revised = NULL,
  document = NULL,
  metadata = NULL,
  score = NULL,
  q = NULL,
  api_key = attackerkb_api_key()
)

kb_assessment(assessment_id, api_key = attackerkb_api_key())
}
\arguments{
\item{assessment_id}{UUID of a specific assessment to return}

\item{editor_id}{UUID of a contributor.}

\item{topid_id}{UUID of the topic this assessment was based on.}

\item{created}{Return all assessments that were created on the given date.}

\item{revised}{Return all assessments that were revised on the given date.}

\item{document}{Text to query the document attribute. A substring match is performed}

\item{metadata}{Text to query the metadata attribute. A substring match is performed}

\item{score}{Return all assessments with this score.}

\item{q}{Return all assessments that have content that match the query string.}

\item{api_key}{See \code{\link[=attackerkb_api_key]{attackerkb_api_key()}}}
}
\description{
The main \code{kb_contributors} function maps 1:1 to the API. Leave values
\code{NULL} that you do not want included in the search parameters.
}
\details{
Set the option \code{progress_enabled} to \code{FALSE} to disable progress spinner.
}
\references{
\url{https://api.attackerkb.com/api-docs/docs}
}
