% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor.R
\name{kb_contributors}
\alias{kb_contributors}
\alias{kd_contributor}
\title{Helpers to query AttackerKB contributors}
\usage{
kb_contributors(
  contributor_id = NULL,
  username = NULL,
  avatar = NULL,
  created = NULL,
  score = NULL,
  q = NULL,
  api_key = attackerkb_api_key()
)

kd_contributor(contributor_id, api_key = attackerkb_api_key())
}
\arguments{
\item{contributor_id}{UUID of a specific contributror to return}

\item{username}{Return contributors with the matching username.}

\item{avatar}{Return all contributors where avatar matches the given value}

\item{created}{Return all contributors that were created on the given date.}

\item{score}{Return all contributors with this score.}

\item{q}{Return all contributors that have usernames that match the query string.}

\item{api_key}{See \code{\link[=attackerkb_api_key]{attackerkb_api_key()}}}
}
\description{
The main \code{kb_contributors} function maps 1:1 to the API. Leave values
\code{NULL} that you do not want included in the search parameters.
}
\references{
\url{https://api.attackerkb.com/api-docs/docs}
}
