% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{kb_topics}
\alias{kb_topics}
\alias{kb_topic}
\title{Helpers to query AttackerKB topics}
\usage{
kb_topics(
  topic_id = NULL,
  editor_id = NULL,
  name = NULL,
  created = NULL,
  revised = NULL,
  disclosed = NULL,
  document = NULL,
  metadata = NULL,
  featured = NULL,
  q = NULL,
  api_key = attackerkb_api_key()
)

kb_topic(
  topic_id = "131226a6-a1e9-48a1-a5d0-ac94baf8dfd2",
  api_key = attackerkb_api_key()
)
}
\arguments{
\item{topic_id}{UUID of a specific topic to return}

\item{editor_id}{UUID of a contributor}

\item{name}{Text to query the name attribute. A substring match is performed}

\item{created}{Return all topics that were created on the given date.}

\item{revised}{Return all topics that were revised on the given date.}

\item{disclosed}{Return all topics that were disclosed on the given date.}

\item{document}{Text to query the document attribute. A substring match is performed}

\item{metadata}{Text to query the metadata attribute. A substring match is performed}

\item{featured}{(lgl) \code{TRUE}/\code{FALSE}. Return all topics that are featured.}

\item{q}{Return all topics that have content that matches the query string.}

\item{api_key}{See \code{\link[=attackerkb_api_key]{attackerkb_api_key()}}}
}
\description{
The main \code{kb_topics} function maps 1:1 to the API. Leave values
\code{NULL} that you do not want included in the search parameters.
}
\details{
Set the option \code{progress_enabled} to \code{FALSE} to disable progress spinner.
}
\references{
\url{https://api.attackerkb.com/api-docs/docs}
}
