% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attackerkb-package.R
\docType{package}
\name{attackerkb}
\alias{attackerkb}
\alias{attackerkb-package}
\title{Tools to Query the Rapid7 AttackerKB API}
\description{
Rapid7 manages a service — \url{https://attackerkb.com/} - where experts can evaluate
various aspects of emergent or existing vulnerabilities and the community can query and
retrieve results. Tools are provided to query the AttackerKB API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/attackerkb}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/attackerkb/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
